server/pyproject.toml
[tool.poetry]
name = "server"
version = "0.1.0"
description = ""
authors = ["wigya <amajewski505@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.5"
flask = "2.3.3"
requests = "^2.31.0"
psycopg2-binary = "^2.9.7"
flask-cors = "^4.0.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
